refactor: Move test cases from test_api to test_validations
